2024/src/day5/day5.test.ts
import { part1, part2 } from "./day5.ts";
import { assertEquals } from "@std/assert";
import { join } from "@std/path";
import { getLines } from "#utils/string.ts";
import { describe, it } from "@std/testing/bdd";

describe("Day5", () => {
  it("should work for part 1 with simple input", async () => {
    const simplePath = join(import.meta.dirname!, "simple-input.txt");
    const inputText = await Deno.readTextFile(simplePath);
    const lines = getLines(inputText);

    const result = part1(lines);

    assertEquals(result, 143);
  });

  it("should work for part 1 with real input", async () => {
    const simplePath = join(import.meta.dirname!, "input.txt");
    const inputText = await Deno.readTextFile(simplePath);
    const lines = getLines(inputText);

    const result = part1(lines);

    assertEquals(result, 4905);
  });
  it("should work for part 2 with simple input", async () => {
    const simplePath = join(import.meta.dirname!, "simple-input.txt");
    const inputText = await Deno.readTextFile(simplePath);
    const lines = getLines(inputText);

    const result = part2(lines);

    assertEquals(result, 123);
  });
  it("should work for part 2 with real input", async () => {
    const simplePath = join(import.meta.dirname!, "input.txt");
    const inputText = await Deno.readTextFile(simplePath);
    const lines = getLines(inputText);

    const result = part2(lines);

    assertEquals(result, 6204);
  });
});
